Just add everything into the installer
Instead of maintaining a list ofr toplevel directories and files
that get added to the installer add everything.

Both the filesystem and the rebase package emit a warning on upgrade
because the list of path sto include wasn't updated. I don't see a reason why
we should exclude anything, so don't.

Fixes #10