Use WizardStyle=Modern instead of Aero
The later seems to be the default for some reason but does not
support dark mode (or breaks when dark mode is enabled).
Use the "Modern" style instead, and also set a smaller default
width while at it.

Note that the style is a QWizard concept and not QTIFW specific.

The only related upstream issue I found was
https://bugreports.qt.io/browse/QTBUG-123853

Fixes #84